change typos
