+++ src/screens/History.js
@@ -1,10 +1,22 @@
 import React from 'react';
 import { View, Text } from 'react-native';
 
-export default function History() {
+export default function NewsDetail() {
   return (
-    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-      <Text>Reading History</Text>
+    <View style={styles.container}>
+      <Text style={styles.text}>Reading History Screen</Text>
     </View>
   );
 }
+
+const styles = {
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  text: {
+    fontSize: 20,
+    fontWeight: 'bold',
+  },
+};
